+++ letsencrypt/accounts/localhost_14000/russtone@yandex.ru/keys/russtone@yandex.ru.key
@@ -1,0 +1,6 @@
+-----BEGIN EC PRIVATE KEY-----
+MIGkAgEBBDB5TgzYmZpoy5oScB2SerOsNuOQ5ZGy3W28tLamENSUaWLhs0KDwyD5
+r8pF/oJr5FygBwYFK4EEACKhZANiAARk0SdSFUqLsG2WxLslHCMT37kqW2QdpRTC
+uEFdcj6493uZp4b1iZ2ZH+eXzmNJhZ8kjocISpsxmL5J/sqlwvm0jl8X+ZuuP6bW
+OUZ3vGbnom0CD6F3t2pici3u4IYc5eQ=
+-----END EC PRIVATE KEY-----
