+++ letsencrypt/certificates/sonar.local.json
@@ -1,0 +1,5 @@
+{
+	"domain": "sonar.local",
+	"certUrl": "https://localhost:14000/certZ/46af3659308b1a26",
+	"certStableUrl": "https://localhost:14000/certZ/46af3659308b1a26"
+}
