+++ letsencrypt/certificates/sonar.local.key
@@ -1,0 +1,6 @@
+-----BEGIN EC PRIVATE KEY-----
+MIGkAgEBBDBse1XPRbFdlE/yGbU2/I65X1tZYmLEzFQjfVhGHHswQIr69IcVtRZX
+8nOr8YZXfRagBwYFK4EEACKhZANiAAT3HfcvjNteF9ac1H84ziASO27nebqDitrK
+X00iJ3HhSi7DdA7NU1quyOMy7FCZWLzgJJ14WaJUwqoOo+vJ6To4oYRHilosIvd6
+dryJETGP+pJ3WPmmvoywirdqlAJTDLg=
+-----END EC PRIVATE KEY-----
