+++ letsencrypt/certificates/sonar.local.pem
@@ -1,0 +1,43 @@
+-----BEGIN CERTIFICATE-----
+MIICujCCAaKgAwIBAgIIRq82WTCLGiYwDQYJKoZIhvcNAQELBQAwKDEmMCQGA1UE
+AxMdUGViYmxlIEludGVybWVkaWF0ZSBDQSAwODJmMjMwHhcNMjAwMzI1MDkxMzU4
+WhcNMjUwMzI1MDkxMzU4WjAWMRQwEgYDVQQDEwtzb25hci5sb2NhbDB2MBAGByqG
+SM49AgEGBSuBBAAiA2IABPcd9y+M214X1pzUfzjOIBI7bud5uoOK2spfTSInceFK
+LsN0Ds1TWq7I4zLsUJlYvOAknXhZolTCqg6j68npOjihhEeKWiwi93p2vIkRMY/6
+kndY+aa+jLCKt2qUAlMMuKOBpzCBpDAOBgNVHQ8BAf8EBAMCBaAwHQYDVR0lBBYw
+FAYIKwYBBQUHAwEGCCsGAQUFBwMCMAwGA1UdEwEB/wQCMAAwHQYDVR0OBBYEFBKX
+UPRxTVZOBDu4I5wSSW7yrt/eMB8GA1UdIwQYMBaAFPMZhC78eCQYhqkwQ4kECHS/
+AYqeMCUGA1UdEQQeMByCC3NvbmFyLmxvY2Fsgg0qLnNvbmFyLmxvY2FsMA0GCSqG
+SIb3DQEBCwUAA4IBAQCqqY7dSOtS7dhdO71Hw5N9W8Ffb3+GSpnm+U6bSuq/6xrW
+w7aExSpeQgO0twtloum13dXzLa+PgycE58rI6yzZG4GTuIoo8JC2vwEecyCkocME
+XJ5Px/Ah2QDedLUjNJPHoaD7lMaLSgxGTwMYsiQFSFFAReNbRxBQe7MpNZ0fi3bM
+nKb8F7FPjIETdM+PXiFwNvrueCCqDizjbOo4s2P/xvZSt6iqI/KlcrBJDjULoWpw
+SPRjbtypzxNXnKKdbS/uzZesQM6wT0/w4xaGrBCGN/NHlPgXHeqVlOj0LneG82qR
+nkeoQElCyX6eWlYspp6a7WPccR0lyzdjF6r6u+ba
+-----END CERTIFICATE-----
+-----BEGIN CERTIFICATE-----
+MIIDUDCCAjigAwIBAgIIajKPZrJW8lMwDQYJKoZIhvcNAQELBQAwIDEeMBwGA1UE
+AxMVUGViYmxlIFJvb3QgQ0EgNGUzNTU2MCAXDTIwMDMyNTA4NTYyNVoYDzIwNTAw
+MzI1MDg1NjI1WjAoMSYwJAYDVQQDEx1QZWJibGUgSW50ZXJtZWRpYXRlIENBIDA4
+MmYyMzCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAMXt/sebluoPxosv
+/ZoA/eOaIut7EzvJJXSqteR6j9j/I1Bb1gzhKvIJ9IDG07VR1B468SowwGT2lGu3
+x4e4w+8LD6o4k0kMNXP2FAWqRE4u8DcN1sAxJZYmSvvCB7Q8N6VCk9wO/v98YNPW
+D6QzXVHHqk5zy+Ra15BsbJ6FaXAfwIxkxpH3bkCRHhF3LEmuFc0tZZu+aHNrgftB
+S5Pgd2H1mbHJga2y5M+M8I8v6WzqaMvoc7iiNb/BIpiq1iTBos7h7BOF/qsWXbsB
+cWkkOgULQhICmNQlrRDDpRr6iSAeYdo8eJnF4y5vNmbbfYX39OqqC86MqTxjKqgY
+q6A4/NMCAwEAAaOBgzCBgDAOBgNVHQ8BAf8EBAMCAoQwHQYDVR0lBBYwFAYIKwYB
+BQUHAwEGCCsGAQUFBwMCMA8GA1UdEwEB/wQFMAMBAf8wHQYDVR0OBBYEFPMZhC78
+eCQYhqkwQ4kECHS/AYqeMB8GA1UdIwQYMBaAFKYUy5uyijvMP52CrQlgGrtgCZiy
+MA0GCSqGSIb3DQEBCwUAA4IBAQCqjLQDiHB7cHkjTJlmXphhNB9Ag69uk5HPaoMF
+gY6seMNCdSEcAGSEkEpuDHnEBvZbEXbfRSeY+cO3eKWaoyxFWTCwQ/WgJZ5QCHrZ
+t17PlEjIcW2hIJfCQOefjDqrd8/RJxL4gJvDH8Y5gsNpL3thDEHO/sHrgnMP6b88
+HjEqRSHNQG2v2haqGGHk3zBYUyfJYFRheXFKTAxV3a47l0LNsEZWddaQEXQqGdqK
+pvKIloUR9NUfb9cC7g3q4+kYeVeMSeu6ARQaL123Y6ffIa6Y+7NY6RV/v97omyT1
+nzfxxyVoAd+LzRNwxynRX+nSAhLwG/sIeTRAiDBFgdlZFbbJ
+-----END CERTIFICATE-----
+-----BEGIN EC PRIVATE KEY-----
+MIGkAgEBBDBse1XPRbFdlE/yGbU2/I65X1tZYmLEzFQjfVhGHHswQIr69IcVtRZX
+8nOr8YZXfRagBwYFK4EEACKhZANiAAT3HfcvjNteF9ac1H84ziASO27nebqDitrK
+X00iJ3HhSi7DdA7NU1quyOMy7FCZWLzgJJ14WaJUwqoOo+vJ6To4oYRHilosIvd6
+dryJETGP+pJ3WPmmvoywirdqlAJTDLg=
+-----END EC PRIVATE KEY-----
